docker/python/app/requirements.txt
python-docx==0.8.11
zipfile36==0.1.3
pymysql==1.0.2
requests==2.27.1
flask==2.0.2
gunicorn==20.1.0
